Updated UI, renamed objects, and enhanced tab switching logic
Significant updates have been made to the RegistrationPortal.Designer.vb, RegistrationPortal.resx, and RegistrationPortal.vb files. The changes primarily focus on improving the user interface, enhancing code readability, and updating the tab switching logic.

In the RegistrationPortal.Designer.vb file, several objects were renamed to more descriptive names, improving code readability and maintainability. The user interface was enhanced with the addition of new objects and the order of some objects was changed. The text displayed on some buttons was also updated for better user experience.

The RegistrationPortal.resx file was updated to include new metadata for several user added columns, indicating that new columns have been added to a data structure in the application.

In the RegistrationPortal.vb file, the System.Diagnostics.Eventing.Reader namespace was imported, and the tab switching logic in several methods was updated. A large block of code was added to the Guna2Button3_Click method, which now retrieves and displays user input. A new method, Guna2Button2_Click_1, was added to enable tab switching to the address tab.